Move gogs (git webview) to an own vhost
Since the .php (PHP) and .md (Markdown) regexes conflict with the gogs
software, gogs was moved to an own vhost.

Files with .php (or .md) at their end got matched also in the /git/
subdirectory which was served by gogs. Nginx throwed a 404 error because
these files obviously did not exist at this place of the directory structure.